Add warning notes to both build runners
Our build necessarily must be run by two different runners (travis and
make) because to minimize warnings, we use the "-W" flag, which turns
warnings into errors. For local development, this is disadvantagous
because sphinx-build fails on the first error, but for travis builds,
not setting the flag means code with warnings can be merged/deployed.

Since the builds must be different, they can't both use the same script
to unify options. These comments are intended to mitigate the risk of
the options diverging over time.