Stop requiring a global installation of grunt
NPM run-scripts use a modified PATH with path/to/project/node_modules/.bin at the start.
This means that running `grunt` in a run-script will use the locally installed grunt instead
of the global one.
This means there won't be any funky issues if you're using a different global version of grunt,
and it removes a global dependency.